requirements.txt
streamlit
langchain
langchain-openai
faiss-cpu
numpy
python-dotenv